Carry "a" flag around when looping back to top of buffer
In an utterly bizarre VimL gotcha, the `&` last flags argument has to go away at
the end because otherwise Vim will use the flags from inside the `if l:last ==#
'a'` loop body, even if the body wasn't evaluated. This is presumably because it
really is evaluated (everything is "eval"), except without side-effects, other
than clobbering the last-flags state, that is.

Amazing.