Work in VISUAL mode
This mostly works pretty well. One gotcha is that `<cword>` expands to the first
word of the selection in VISUAL mode, not the word under the cursor. I think
this is technically because VIM actually does move the cursor to the top before
calling any function.

There may be a workaround, but it will be pretty involved, so I will save it for
another commit (IFF I can get it working).