Take into account 'gdefault' setting
If the user has 'gdefault' set (not the default), then the "g" flag
actually turns *off* global replacement instead of turning it on. So,
check for 'gdefault' in order to decide whether or not to pass the "g"
(we always want to do global replacements).

Reported here:

  https://twitter.com/briemens/status/1080554857998757888

My guess at problem here:

  https://twitter.com/wincent/status/1080616775681798144

Confirmed that it was here:

  https://twitter.com/briemens/status/1080618527218978816