refactor: replace hard-coded "$" with "very magic" list
As the PR duly notes, it's hard to imagine an obvious use case where
some "special" character other than "$" is part of a Vim "word", but if
we're going to do this, we might as well work for all possible
definitions of `'iskeyword'`.

I was originally thinking we should do this based on what is in
`'iskeyword'`, but we can just do a blanket rule and escape literally
everything that might have "very magic" significance. Note that in
preparing the list of characters to be escaped, we have to double escape
the following to produce a valid `:execute` string:

- '"' because the list itself is double-quoted.
- '\' because it is in the list and we need to preserve it.
- '|' because otherwise it will terminate the command.

Test plan:

Set a bizarre `'iskeyword'` like:

    :set iskeyword=@,34,39,41,47-57,92,_,124,192-255

ie. in additional to the typical 48-57,192-255 range, also add:

    34 = "
    39 = '
    41 = )
    47 = /
    92 = \
    124 = |

and then hit the mapping with the cursor over a "word" like this one:

    '"/\|')

(That example chosen simply because it appears in the source.)